pyproject.toml
[project]
name = "barkprints"
version = "0.1.0"
description = "Generate poetry and text from tree bark images using deterministic feature mapping"
authors = [
    {name = "Martin", email = "martin@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "sentence-transformers>=2.2.0",
]

[project.scripts]
barkprints = "barkprints.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/barkprints"]

